Simplify Country/Region/ConsoleRegion to byte from int
Only exists in gen6/7 games.

+++ PKHeX.Core/Legality/Encounters/EncounterMisc/EncounterEgg.cs
@@ -97,7 +97,7 @@ namespace PKHeX.Core
                     break;
                 case (int)Core.Species.Scatterbug or (int)Core.Species.Spewpa or (int)Core.Species.Vivillon:
                     if (sav is IRegionOrigin o)
-                        pk.Form = Vivillon3DS.GetPattern((byte)o.Country, (byte)o.Region);
+                        pk.Form = Vivillon3DS.GetPattern(o.Country, o.Region);
                     // else 0
                     break;
             }
