+++ PKHeX.Core/PKM/Shared/IRegionOrigin.cs
@@ -4,11 +4,11 @@
     {
         /// <summary> Console hardware region. </summary>
         /// <see cref="RegionID"/>
-        int ConsoleRegion { get; set; }
+        byte ConsoleRegion { get; set; }
         /// <summary> Console's configured Country via System Settings. </summary>
-        int Country { get; set; }
+        byte Country { get; set; }
         /// <summary> Console's configured Region within <see cref="Country"/> via System Settings. </summary>
-        int Region { get; set; }
+        byte Region { get; set; }
     }
 
     public static partial class Extensions
